+++ src/components/achievements/index.js
@@ -1,0 +1,3 @@
+import "./achievements.css";
+
+export { achievementsState, achievementsMethods } from "./achievements.js";